refactor(usage): split handler unit tests into own file to satisfy size guard
server/src/server/usage_analytics.rs grew past the 1500-line size-guard
limit after adding the list-price KPI card and its test coverage. Move the
inline #[cfg(test)] mod tests into usage_analytics_handler_tests.rs via
#[path], matching the existing usage_analytics_schema_tests.rs pattern.
No behavior change.
